Spawn sizes to ushort

+++ VolatileHordes/Core/Services/Commands.cs
@@ -35,7 +35,7 @@ namespace VolatileHordes
                 }
                 case "wander":
                 {
-                    if (paramList.Count > 1 && int.TryParse(paramList[1], out var size))
+                    if (paramList.Count > 1 && ushort.TryParse(paramList[1], out var size))
                     {
                         Logger.Info("Artificially spawning a wandering horde of size {0}", size);
                         await Container.WanderingHordeSpawner.Spawn(size);
